refactor(frontend): the company form leaves the screen it was living in
`organizations.tsx` was 2,925 lines and a third of what it exported was not a
screen: the create body, the update body, the address mapping and the field
list that collects them. `companyheader.tsx` builds the same edit from the same
field list, and the only way to reach any of it was to import from that screen
— which is how a form-mapping layer became a screen's public surface.

The form moves to `companyform.ts`, which renders nothing and stays `.ts` so it
cannot drift back into one. The screen drops to 2,589 lines and sheds four
imports it no longer needs.

Both waivers are re-frozen DOWNWARD, which is the direction the ratchet is for:
2,925 to 2,589 here, and 975 to 974 for the header. Leaving the old numbers
would have left the pair 337 lines of slack to grow back into silently.

The overlay-refusal census names the file each request lives in, so the create
entry follows it — the same note the merge entry already carries from the last
time this screen shed a piece.
